requirements.txt
flask
firebase-admin
python-dotenv
flask-session
werkzeug